+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Search skills (combat tab only) and spells
 - Add attainments section to magic tab
+- Add forces spells to compendium
+
+### Changed
+
+- Spell addons can also cost reach
 
 ## [1.7.0] - 2024-04-17