+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add attainments section to magic tab
 - Add forces spells to compendium
 - Add Space: Everywhere Attainment support to spellcasting form
+- Add Time in a Bottle Attainment support to spellcasting form
 
 ### Changed