+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.8.0] - 2024-05-01
+
 ### Added
 
 - Search skills (combat tab only) and spells
@@ -155,7 +157,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Casting spells from modal shows details in chat and updates spell slots
 - Prepare spells as arcanist
 
-[unreleased]: https://github.com/sumayyah-k/swnr-space-magic/compare/1.7.0...HEAD
+[unreleased]: https://github.com/sumayyah-k/swnr-space-magic/compare/1.8.0...HEAD
+[1.8.0]: https://github.com/sumayyah-k/swnr-space-magic/compare/1.7.0...1.8.0
 [1.7.0]: https://github.com/sumayyah-k/swnr-space-magic/compare/1.6.1...1.7.0
 [1.6.1]: https://github.com/sumayyah-k/swnr-space-magic/compare/1.6.0...1.6.1
 [1.6.0]: https://github.com/sumayyah-k/swnr-space-magic/compare/1.5.3...1.6.0